Fix formatting issues in stdout (with int64 bit integers) + move H2T_loadMesh fonction and create fortran APIs.
