+++ src/hashhexa.c
@@ -243,7 +243,7 @@ int H2T_edgePut(pHedge hash,int a,int b,int np) {
     ha      = &hash->item[hash->hnxt];
     ++hash->hnxt;
     if ( hash->hnxt >= hash->nhmax ) {
-      fprintf(stdout,"  ## Memory alloc problem (edge): %d\n",hash->nhmax);
+      fprintf(stdout,"  ## Memory alloc problem (edge): %" MMG5_PRId "\n",hash->nhmax);
       return 0;
     }
   }
